feat: refactor stepzenProjectScanner into modular service architecture
Split the monolithic stepzenProjectScanner.ts (743 LoC) into focused,
testable modules while maintaining 100% backward compatibility.

## New Architecture

### Service Layer
- services/SchemaIndexService.ts: Main orchestration service
- services/index.ts: Updated service registry with schemaIndex

### Schema Processing Layer
- services/schema/indexer.ts: Core indexing logic and data structures
- services/schema/linker.ts: Loop-safe SDL include traversal
- services/schema/parser.ts: GraphQL AST utility functions

### Compatibility Layer
- utils/stepzenProjectScanner.ts: Backward-compatible API wrapper

## Benefits
- Improved modularity with single-responsibility modules
- Enhanced testability through dependency injection
- Better maintainability with clear separation of concerns
- Preserved performance with optimized data structures

## Testing
- All 101 tests passing (including STRUCT1-TESTS)
- Service registry tests updated for new schemaIndex service
- Full backward compatibility verified

## Documentation
- Added comprehensive architecture documentation
- Detailed migration strategy and benefits
- Clear module responsibilities and data flow

Resolves: #49
