+++ src/services/index.ts
@@ -1,6 +1,7 @@
 import { StepzenCliService } from './cli';
 import { Logger, logger } from './logger';
 import { ProjectResolver } from './projectResolver';
+import { SchemaIndexService } from './SchemaIndexService';
 
 /**
  * Service registry for dependency injection of application services
@@ -10,6 +11,7 @@ export interface ServiceRegistry {
   cli: StepzenCliService;
   logger: Logger;
   projectResolver: ProjectResolver;
+  schemaIndex: SchemaIndexService;
 }
 
 /**
@@ -19,6 +21,7 @@ export const services: ServiceRegistry = {
   cli: new StepzenCliService(),
   logger,
   projectResolver: new ProjectResolver(logger),
+  schemaIndex: new SchemaIndexService(),
 };
 
 /**
